move select_input/output to static/controw flow
