+++ test/legacy_test/test_select_input_output_op.py
@@ -22,7 +22,7 @@ from paddle.fluid import core
 from paddle.fluid.backward import append_backward
 from paddle.fluid.executor import Executor
 from paddle.fluid.framework import Program, program_guard
-from paddle.fluid.layers.control_flow import select_input, select_output
+from paddle.static.nn.control_flow import select_input, select_output
 
 paddle.enable_static()
